backend/pyproject.toml
[project]
name = "mlexperience-backend"
version = "1.0.0"
description = "ML Experience System backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0",
    "pandas>=2.0",
    "numpy>=2.0",
    "scikit-learn>=1.5",
    "torch>=2.5",
    "pytorch-lightning>=2.4",
    "torchmetrics>=1.0",
    "psutil>=5.9",
    "ray[tune]>=2.40",
    "optuna>=3.0",
    "psycopg[binary]>=3.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
    "pytest>=8.0",
    "httpx>=0.27",  # FastAPI TestClient dependency
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]


[tool.ruff]
target-version = "py312"
line-length = 130
src = ["."]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
